consistency

+++ src/_ext.cpp
@@ -373,7 +373,7 @@ PYBIND11_MODULE(_ext, m) {
         }
         jk++;
 
-        auto parid = py::array(py::buffer_info(nullptr, sizeof(int64_t), py::format_descriptor<int>::value, 1, {sizepar}, {sizeof(int64_t)}));
+        auto parid = py::array(py::buffer_info(nullptr, sizeof(int64_t), py::format_descriptor<int64_t>::value, 1, {sizepar}, {sizeof(int64_t)}));
         auto bufparid = parid.request();
         int64_t *ptrid = (int64_t *)bufparid.ptr;
